+++ Data_Analysis/LammpsConstants.m
@@ -11,7 +11,7 @@ function [varargout] = LammpsConstants()
 
 %% Squeeze Physical constants, energy converter, mass converter, length conveter, time converter
 
-constants_name          =   ['k_b n_a kcal2j kcalm2j kcalm2t g2kg gm2kg an2m nm2m fs2s ps2s'];
+constants_name          =   ['k_b n_a kcal2j kcalm2j kcalm2t g2kg gm2kg an2m an2nm nm2m fs2s ps2s ns2s'];
 constants_name          =   split(constants_name);
                         
 constants_num           =   length(constants_name);
@@ -24,9 +24,11 @@ kcalm2t                     =   kcalm2j/k_b;
 g2kg                        =   1e-3;
 gm2kg                       =   g2kg/n_a;
 an2m                        =   1e-10;
+an2nm                       =   1e-1;
 nm2m                        =   1e-9;
 fs2s                        =   1e-15;
 ps2s                        =   1e-12;
+ns2s                        =   1e-9;
 
 %% -----------------------Output-----------------------